chore: fix new clippy warnings on nightly
All changes made using `cargo clippy --fix`.

Great to see old code taking full advantage of let-chains, with the
toolchain doing all the work for us :)

+++ holo-ldp/src/collections.rs
@@ -340,15 +340,14 @@ impl Adjacencies {
                 o.remove_entry();
             }
         }
-        if let Some(ifname) = &adj.source.ifname {
-            if let hash_map::Entry::Occupied(mut o) =
+        if let Some(ifname) = &adj.source.ifname
+            && let hash_map::Entry::Occupied(mut o) =
                 self.iface_tree.entry(ifname.clone())
-            {
-                let tree = o.get_mut();
-                tree.remove(&adj.source);
-                if tree.is_empty() {
-                    o.remove_entry();
-                }
+        {
+            let tree = o.get_mut();
+            tree.remove(&adj.source);
+            if tree.is_empty() {
+                o.remove_entry();
             }
         }
